Commented all of the codes and fixed bugs

+++ SProtector/Module1.vb
@@ -1,8 +1,11 @@
 ﻿Imports dnlib.DotNet
-
+'SProtector by sahandevs
+'Github : https://github.com/sahand100/
 
 
 Module Module1
+
+    'Log with a blue bar back of text for Info messages
     Sub LogInfo(ByVal text As String)
         Dim k = Console.BackgroundColor
         Console.BackgroundColor = ConsoleColor.Blue
@@ -10,6 +13,7 @@ Module Module1
         Console.BackgroundColor = k
         Console.WriteLine(" " + text)
     End Sub
+    'Log with a red bar back of text for Error messages
     Sub LogError(ByVal text As String)
         Dim k = Console.BackgroundColor
         Console.BackgroundColor = ConsoleColor.Red
@@ -17,9 +21,11 @@ Module Module1
         Console.BackgroundColor = k
         Console.WriteLine(" " + text)
     End Sub
+
+    'Log with a green bar back of text for good messages
     Sub LogGood(ByVal text As String)
         Dim k = Console.BackgroundColor
-        Console.BackgroundColor = ConsoleColor.Green
+        Console.BackgroundColor = ConsoleColor.DarkGreen
         Console.Write("   ")
         Console.BackgroundColor = k
         Console.WriteLine(" " + text)
@@ -30,75 +36,71 @@ Module Module1
 #If DEBUG Then
             input = "1.exe"
 #Else
- input = sargs(0).ToString()
+            input = sargs(0).ToString() 'read file path from first arg
 #End If
         Catch ex As Exception
-            Console.WriteLine("Please Drag & Drop A File")
+            Console.WriteLine("Please Drag & Drop A File") 'show this message if something goes wrong
             w()
             Return
         End Try
         Try
 
-            SProtector.GlobAssembly.AsmInputPath = input
-            If input.Contains(".exe") Then
-                SProtector.GlobAssembly.AsmOutputPath = input.Replace(".exe", "").Replace(".dll", "") + "-sprotected.exe"
+            SProtector.GlobAssembly.AsmInputPath = input 'set my global var " assembly path " to first arg
+            If input.Contains(".exe") Then 'Check if it is a exe file
+                SProtector.GlobAssembly.AsmOutputPath = input.Replace(".exe", "").Replace(".dll", "") + "-sprotected.exe" ' set my global var " assembly output path " to first arg + -protected.exe if file is exe file
             Else
-                SProtector.GlobAssembly.AsmOutputPath = input.Replace(".exe", "").Replace(".dll", "") + "-sprotected.dll"
+                SProtector.GlobAssembly.AsmOutputPath = input.Replace(".exe", "").Replace(".dll", "") + "-sprotected.dll" ' set my global var " assembly output path " to first arg + -protected.dll if file is dll file
             End If
-            SProtector.GlobAssembly.Asm2 = AssemblyDef.Load("Methods.exe")
-            SProtector.GlobAssembly.Asm = AssemblyDef.Load(SProtector.GlobAssembly.AsmInputPath)
+            SProtector.GlobAssembly.Asm2 = AssemblyDef.Load("Methods.exe") ' loads external exe files for injecting methods to my global var "asm2"
+            SProtector.GlobAssembly.Asm = AssemblyDef.Load(SProtector.GlobAssembly.AsmInputPath) ' load main exe file for protecting to my global var "asm"
 
-            'LogInfo("Protecting With [Rex Anti Debug]...")
-            'SProtector.Protectors.AntiDebug.Protect()
-            'LogGood("Protecting With [Rex Anti Debug](Done)")
+            ' Not by now you cant use both constants or crasher and antidebug together
 
+            LogInfo("Protecting With [Rex Anti Debug]...")
+            SProtector.Protectors.AntiDebug.Protect() ' Protect with anti debug method
+            LogGood("Protecting With [Rex Anti Debug](Done)")
+
+            SProtector.GlobAssembly.Asm2 = AssemblyDef.Load("Methods.exe") ' reloads external exe files for injecting methods to my global var "asm2"
             LogInfo("Protecting With [Constants]...")
-            SProtector.Protectors.Constants.Protect()
+            SProtector.Protectors.Constants.Protect() ' Protect with constants ( Crypting strings and integers ) method
             LogGood("Protecting With [Constants](Done)")
 
-            LogInfo("Protecting With [Crasher]...")
-            SProtector.Protectors.Constants.Protect()
-            LogGood("Protecting With [Crasher](Done)")
+            'LogInfo("Protecting With [Crasher]...")
+            'SProtector.Protectors.Constants.Protect() 'i've removed this but i'll add it soon
+            'LogGood("Protecting With [Crasher](Done)")
 
 
 
             LogInfo("Protecting With [AntiILDasm]...")
-            SProtector.Protectors.AntiILDasm.Protect()
+            SProtector.Protectors.AntiILDasm.Protect() ' Protect with AntiIlDasm with confuserex method
             LogGood("Protecting With [AntiILDasm](Done)")
 
             LogInfo("Protecting With [FakeModules]...")
-            SProtector.Protectors.FakeModules.Protect()
+            SProtector.Protectors.FakeModules.Protect() ' Protect file by adding a lot of fake modules
             LogGood("Protecting With [FakeModules](Done)")
 
-
-
-
-
             LogInfo("Protecting With [Renamer]...")
-            SProtector.Protectors.Renamer.Protect()
+            SProtector.Protectors.Renamer.Protect() 'Protect method , fields and ... by changing their name to a random string
             LogGood("Protecting With [Renamer](Done)")
 
 
-            LogInfo("Saved In : " + SProtector.GlobAssembly.AsmOutputPath)
             '   Process.Start(SProtector.GlobAssembly.AsmOutputPath)
+            If My.Computer.FileSystem.FileExists(SProtector.GlobAssembly.AsmOutputPath) Then My.Computer.FileSystem.DeleteFile(SProtector.GlobAssembly.AsmOutputPath)  ' delete output file if exists
 
+            SProtector.GlobAssembly.Asm.Modules(0).Write(SProtector.GlobAssembly.AsmOutputPath) ' save file
 
-            Try
-                My.Computer.FileSystem.DeleteFile(SProtector.GlobAssembly.AsmOutputPath)
-            Catch ex As Exception
-
-            End Try
-
-            SProtector.GlobAssembly.Asm.Modules(0).Write(SProtector.GlobAssembly.AsmOutputPath)
-
+            LogInfo("Saved In : " + SProtector.GlobAssembly.AsmOutputPath)
             w()
         Catch ex As Exception
-            LogError(ex.ToString)
-            '      Console.WriteLine("Error : " + ex.ToString)
+#If DEBUG Then
+             LogError(ex.ToString) ' log fully detail if it is on debug mode
+#Else
+            LogError(ex.Message) 'log error text
+#End If
             w()
         End Try
     End Sub
-    Sub w()
+    Sub w() ' a shortcut for console.readkey
         Console.ReadKey()
     End Sub
 End Module
